backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# Caching & Background Jobs
redis>=4.5.2,<5.0.0
celery[redis]==5.3.4

# External APIs
garminconnect==0.2.13
anthropic==0.7.7
httpx==0.25.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Utilities
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
